+++ src/locatives/class.lisp
@@ -5,7 +5,6 @@
                 #:locate-object
                 #:define-locative-type)
   (:import-from #:40ants-doc/render/args)
-  (:import-from #:40ants-doc/builder/bullet)
   (:import-from #:40ants-doc/reference-api
                 #:canonical-reference)
   (:import-from #:40ants-doc/args)
@@ -39,26 +38,6 @@
   (if (subtypep class 'condition)
       (40ants-doc/reference::make-reference (class-name class) 'condition)
       (40ants-doc/reference::make-reference (class-name class) 'class)))
-
-;; (defmethod document-object ((class class) stream)
-;;   (let* ((conditionp (subtypep class 'condition))
-;;          (symbol (class-name class))
-;;          (superclasses
-;;            (remove-if (lambda (name)
-;;                         (or (eq name 'standard-object)
-;;                             (and conditionp (eq name 'condition))))
-;;                       (mapcar #'class-name
-;;                               (swank-mop:class-direct-superclasses class)))))
-;;     (40ants-doc/builder/bullet::print-bullet class stream)
-;;     (when superclasses
-;;       (write-char #\Space stream)
-;;       (if 40ants-doc/builder/vars::*document-mark-up-signatures*
-;;           (40ants-doc/render/args::print-arglist
-;;            (mark-up-superclasses superclasses) stream)
-;;           (40ants-doc/render/args::print-arglist superclasses stream)))
-;;     (40ants-doc/builder/bullet::print-end-bullet stream)
-;;     (40ants-doc/args::with-dislocated-symbols ((list symbol))
-;;       (40ants-doc/render/print::maybe-print-docstring class t stream))))
 
 
 (defmethod 40ants-doc/commondoc/builder:to-commondoc ((class class))