+++ src/locatives/macro.lisp
@@ -5,7 +5,6 @@
                 #:locate-object
                 #:define-locative-type)
   (:import-from #:40ants-doc/render/args)
-  (:import-from #:40ants-doc/builder/bullet)
   (:import-from #:40ants-doc/reference-api
                 #:canonical-reference)
   (:import-from #:40ants-doc/args)