+++ src/locatives/package.lisp
@@ -5,7 +5,6 @@
                 #:locate-object
                 #:define-locative-type)
   (:import-from #:40ants-doc/render/args)
-  (:import-from #:40ants-doc/builder/bullet)
   (:import-from #:40ants-doc/reference-api
                 #:canonical-reference)
   (:import-from #:40ants-doc/args)
@@ -33,13 +32,6 @@
 
 (defmethod canonical-reference ((package package))
   (40ants-doc/reference::make-reference (package-name package) 'package))
-
-;; (defmethod document-object ((package package) stream)
-;;   (let ((symbol (package-name package)))
-;;     (40ants-doc/builder/bullet::print-bullet package stream)
-;;     (40ants-doc/builder/bullet::print-end-bullet stream)
-;;     (40ants-doc/args::with-dislocated-symbols ((list symbol))
-;;       (40ants-doc/render/print::maybe-print-docstring package t stream))))
 
 
 (defmethod 40ants-doc/commondoc/builder:to-commondoc ((package package))