+++ src/locatives/slots.lisp
@@ -6,7 +6,6 @@
                 #:locate-object
                 #:define-locative-type)
   (:import-from #:40ants-doc/render/args)
-  (:import-from #:40ants-doc/builder/bullet)
   (:import-from #:40ants-doc/reference-api
                 #:canonical-reference)
   (:import-from #:40ants-doc/args)